wip singleton cuinter
