+++ game/common.py
@@ -25,9 +25,9 @@ class DamageInstance(NamedTuple):
 class Stats(NamedTuple):
     """Stores the stats of a charcter.
 
-	Use more than one instance per character to differentiate base and current
-	stats.
-	"""
+    Use more than one instance per character to differentiate base and current
+    stats.
+    """
     max_health: int = None   # if it drops to 0, you die
     max_stamina: int = None  # used as a resource for performing PHYSICAL actions
     max_mana: int = None     # used as a resource for performing MAGICAL actions
@@ -204,3 +204,13 @@ class Character(NamedTuple):
             else:
                 continue
         return f"{self.name} (♥ {self.health})"
+
+
+class DialogLine(NamedTuple):
+    text: str
+    character: Character = None
+
+
+def move_toward(a: int | float, b: int | float, step: int | float = 1) -> int | float:
+    """Returns a moved by step towards b without overshooting."""
+    return min(a + step, b) if b >= a else max(a - step, b)
