+++ game/cuinter.py
@@ -1,27 +1,29 @@
 """
-Système d'affichage d'interface.
+Système de gestion d'interface pour curses.
 
 Contributors:
     Romain
 """
 
-from common import Stats, DamageInstance, Character, Action
+from common import DialogLine, move_toward
 import copy
-import curses
 import ctypes
-import logging
+import curses
+import logger
 import math
 import time
 from typing import NamedTuple
 
 X_CORRECTION = 2.6 # Hauteur / largeur d'un caractère
 CHARACTER_TIME = 0.025 # Délai d'affichage de chaque caractère dans les textes
+EMPTY_BUFFER = [[None for _ in range(screen_width)] for _ in range(screen_height)]
 
+logger = logging.getLogger(__name__)
+logging.basicConfig(filename='cuinter.log', encoding='utf-8', level=logging.DEBUG)
 
-class DialogLine(NamedTuple):
-    text: str
-    character: Character = None
-
+stdscr = None
+screen_height = None
+screen_width = None
 
 class Rectangle(NamedTuple):
     y: int
@@ -32,6 +34,7 @@ class Rectangle(NamedTuple):
     
     @classmethod
     def new(cls, y: int, x: int, height: int, width: int, color: int = 255) -> "Rectangle":
+        logger.debug("Created new Rectangle")
         return cls(y, x, height, width, color)
     
     def config(self, **kwargs) -> "Rectangle":
@@ -86,6 +89,7 @@ class TextBox(NamedTuple):
     
     @classmethod
     def new(cls, y: int, x: int, height: int, width: int, text: str = None) -> "TextBox":
+        logger.debug("Created new TextBox")
         return cls(
             Rectangle(y, x, height, width),
             text,
@@ -139,6 +143,7 @@ class DialogBox(NamedTuple):
     @classmethod
     def new(cls, y: int, x: int, height: int, width: int, dialog: tuple[DialogLine, ...]) -> "DialogBox":
         return cls(
+            logger.debug("Created new DialogBox")
             TextBox.new(y, x, height, width),
             dialog,
             time.time(),
@@ -208,6 +213,7 @@ class ChoiceBox(NamedTuple):
     
     @classmethod
     def new(cls, y: int, x: int, height: int, width: int, options: tuple[str, ...], selected_index: int = 0) -> "ChoiceBox":
+        logger.debug("Created new ChoiceBox")
         return cls(
             TextBox.new(y, x, height, width),
             options,
@@ -249,15 +255,18 @@ class ChoiceBox(NamedTuple):
         return updated_self
 
 
-def move_toward(a: int | float, b: int | float, step: int | float = 1) -> int | float:
-    return min(a + step, b) if b >= a else max(a - step, b)
-
-
 def set_cell(buffer: list[list[tuple[str, int]]], y: int, x: int, char: str = "█", color: int = 255) -> None:
     """Remplace la cellule à la position y x du buffer sans erreurs d'index."""
     if 0 <= y < len(buffer) and 0 <= x < len(buffer[0]):
         buffer[y][x] = (char, color) if not char is None else None
-    
+
+
+def fullscreen() -> None:
+    """Simule la pression de la touche F11."""
+    user32 = ctypes.windll.user32
+    user32.keybd_event(0x7A, 0, 0, 0)
+    user32.keybd_event(0x7A, 0, 0x0002, 0)
+
 
 def display_buffer(stdscr, buffer: tuple[tuple[tuple[str, int]]]) -> None:
     """
@@ -288,112 +297,19 @@ def display_buffer(stdscr, buffer: tuple[tuple[tuple[str, int]]]) -> None:
         stdscr.addstr(row_str)
 
 
-def fullscreen() -> None:
-    """Simule la pression de la touche F11."""
-    user32 = ctypes.windll.user32
-    user32.keybd_event(0x7A, 0, 0, 0)
-    user32.keybd_event(0x7A, 0, 0x0002, 0)
+def start() -> None:
+    time.sleep(0.5)
+    fullscreen()
+    time.sleep(0.5)
+    curses.wrapper(_main)
 
 
-def main(stdscr) -> None:
-    # Initialisation couleurs curses
-    curses.start_color()
-    curses.use_default_colors()
-    for i in range(curses.COLOR_PAIRS - 1):
-        curses.init_pair(i + 1, i, -1)
-
-    curses.curs_set(0) # Cache le curseur
-    stdscr.nodelay(1) # Pas de blocage d'entrées
-    stdscr.timeout(0) # Délai de vérification d'entrée
-
-    screen_height, screen_width = stdscr.getmaxyx()
-
-    empty_buffer = [[None for _ in range(screen_width)] for _ in range(screen_height)]
-
-    last_time = time.time()
-    fps_timer = last_time  # Temps de la dernière mise à jour de l'affichage FPS
-    frame_count = 0
-    average_fps = 0
+def update(delta_time: float) -> None:
     
-    dialog = (
-        DialogLine("LELOLELOELOLEOLEOLEOLEOLEOLOLEOLOEELOmmmmmmmmmmmmmmmmmm    yeseiurrrrrhjsdhdjhsdjhsdhjsdhjdshjsdjhsdjhdsjhdshjsdhjsdhjsdhjdshjdshdsdssjhgfqwè¨qè¨¨èwq¨qwèwq", Character("Idris")),
-        DialogLine("bruh"),
-        DialogLine("AAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
-        DialogLine("We're no strangers to love You know the rules and so do I A full commitment's what I'm thinkin' of You wouldn't get this from any other guy I just wanna tell you how I'm feeling Gotta make you understand Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you We've known each other for so long Your heart's been aching, but you're too shy to say it Inside, we both know what's been going on We know the game and we're gonna play it And if you ask me how I'm feeling Don't tell me you're too blind to see Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you We've known each other for so long Your heart's been aching, but you're too shy to say it Inside, we both know what's been going on We know the game and we're gonna play it I just wanna tell you how I'm feeling Gotta make you understand Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you Never gonna give you up, never gonna let you down Never gonna run around and desert you Never gonna make you cry, never gonna say goodbye Never gonna tell a lie and hurt you", Character("Rick Astley")),
-    )
-    dialog_box = DialogBox.new(
-        screen_height,
-        50,
-        10,
-        screen_width - 100,
-        dialog,
-    )
+
+
+def _main(stdscr_instance) -> None:
+    global stdscr
+    stdscr = stdscr_instance
     
-    options = (
-        "haram",
-        "harambe",
-        "PETAH",
-        "The honse is here.",
-    )
-    choice_box = None
     
-    float_dialog_y = float(dialog_box.y)
-
-    while 1:
-        current_time = time.time()
-        delta_time = current_time - last_time # Temps écoulé depuis la dernière frame
-        last_time = current_time
-
-        frame_count += 1
-        if current_time - fps_timer >= 1.0:
-            average_fps = frame_count / (current_time - fps_timer)
-            fps_timer = current_time
-            frame_count = 0
-
-        key = stdscr.getch()
-        if key == ord(" ") or key == ord("\n"):
-            if not dialog_box is None:
-                dialog_box = dialog_box.next()
-                if dialog_box is None:
-                    choice_box = ChoiceBox.new(
-                        screen_height - 12,
-                        50,
-                        10,
-                        screen_width - 100,
-                        options,
-                    )
-        elif key == curses.KEY_UP:
-            if not choice_box is None:
-                choice_box = choice_box.select_previous()
-        elif key == curses.KEY_DOWN:
-            if not choice_box is None:
-                choice_box = choice_box.select_next()
-        elif key == ord("q"):
-            break
-
-        stdscr.clear()
-        
-        buffer = copy.deepcopy(empty_buffer)
-        
-        if not dialog_box is None:
-            dialog_target_y = screen_height if dialog_box.start_time is None else screen_height - 12
-            if dialog_box.y != dialog_target_y:
-                float_dialog_y = move_toward(float_dialog_y, dialog_target_y, delta_time * 100)
-                dialog_box = dialog_box.config(y=round(float_dialog_y))
-            dialog_box = dialog_box.draw(buffer)
-        
-        if not choice_box is None:
-            choice_box = choice_box.draw(buffer)
-        
-        display_buffer(stdscr, buffer)
-        stdscr.addstr(0, 0, f"FPS : {round(average_fps)}")
-
-        stdscr.refresh()
-
-
-time.sleep(0.5)
-fullscreen()
-time.sleep(0.5)
-logger = logging.getLogger(__name__)
-logging.basicConfig(filename='dialog.log', encoding='utf-8', level=logging.DEBUG)
-curses.wrapper(main)
